Create Comparing_Sample_Preparation_Methods_Mixing_vs_NoMixing
This code was specifically used to compare mixing and not mixing samples after a harvest in a qRT-PCR workflow that is becoming automated. However, this code is adaptable for comparing different conditions to each other to look for significant changes. Instead of "Mix" vs "NotMixed" conditions it can be, "Freeze-Dried" vs "Not Freeze-Dried" samples or master mix, etc...